fix work status collection
